split export code btw cbio specific vs export generic
